# verification/golden/contracts/openapi/web-bff.openapi.yaml
openapi: 3.1.0
info:
  title: web-bff
  version: 0.1.0
paths:
...
  /api/user/me:
    get:
      tags:
      - user
      summary: Get current user profile.
      operationId: get_user_profile
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /api/user/tenants:
    get:
      tags:
      - user
      summary: List user's tenant bindings.
      operationId: get_user_tenants
      responses:
        '200':
          description: User tenants retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserTenantResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
...
components:
  schemas:
...
    ErrorCode:
      type: string
      description: |-
        Machine-readable error codes.

        These codes are stable across releases and form the public API contract.
      enum:
      - BadRequest
      - Unauthorized
      - Forbidden
      - NotFound
      - Conflict
      - ValidationError
      - RateLimited
      - InternalError
      - ServiceUnavailable
      - DatabaseError
      - CacheError
      - ExternalServiceError
    ErrorResponse:
      type: object
      description: |-
        Standardized API error response.

        Every error returned by the API conforms to this shape.
        Callers depend on this public DTO, not on individual service error enums.
      required:
      - code
      - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          description: Machine-readable error code (e.g., `validation_error`, `not_found`).
        details:
          description: |-
            Optional additional details for debugging.
            Never sent in production responses.
        message:
          type: string
          description: Human-readable error message.
...
    UserProfileResponse:
      type: object
      description: Current authenticated user profile.
      required:
      - id
      - user_sub
      - display_name
      - created_at
      properties:
        created_at:
          type: string
        display_name:
          type: string
        email:
          type:
          - string
          - 'null'
        id:
          type: string
        last_login_at:
          type:
          - string
          - 'null'
        user_sub:
          type: string
    UserTenantResponse:
      type: object
      description: Tenant binding visible to the authenticated user.
      required:
      - tenant_id
      - role
      - joined_at
      properties:
        joined_at:
          type: string
        role:
          type: string
        tenant_id:
          type: string
        tenant_name:
          type:
          - string
          - 'null'
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer JWT used to build authenticated request context
...